requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging @ file:///home/task_176104874243446/conda-bld/packaging_1761049080023/work
python-dotenv==1.2.2
setuptools==80.10.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
wheel==0.46.3
WTForms==3.2.1

